[derive] Start work on kvp parser macro
