+++ bve/src/parse/kvp/mod.rs
@@ -20,10 +20,13 @@
 //! ```
 
 use crate::parse::Span;
+pub use parse::*;
+
+mod parse;
 
 #[derive(Debug, Clone, Eq, PartialEq)]
 pub struct KVPFile<'s> {
-    pub sections: Vec<Section<'s>>,
+    pub sections: Vec<KVPSection<'s>>,
 }
 
 impl<'s> Default for KVPFile<'s> {
@@ -35,13 +38,13 @@ impl<'s> Default for KVPFile<'s> {
 }
 
 #[derive(Debug, Clone, Eq, PartialEq)]
-pub struct Section<'s> {
+pub struct KVPSection<'s> {
     pub name: Option<&'s str>,
     pub span: Span,
-    pub values: Vec<Value<'s>>,
+    pub values: Vec<KVPValue<'s>>,
 }
 
-impl<'s> Default for Section<'s> {
+impl<'s> Default for KVPSection<'s> {
     fn default() -> Self {
         Self {
             name: None,
@@ -52,7 +55,7 @@ impl<'s> Default for Section<'s> {
 }
 
 #[derive(Debug, Clone, Eq, PartialEq)]
-pub struct Value<'s> {
+pub struct KVPValue<'s> {
     pub span: Span,
     pub data: ValueData<'s>,
 }
@@ -63,203 +66,14 @@ pub enum ValueData<'s> {
     Value { value: &'s str },
 }
 
-#[must_use]
-#[allow(clippy::single_match_else)] // This advises less clear code
-pub fn parse_kvp_file(input: &str) -> KVPFile<'_> {
-    let mut file = KVPFile::default();
-    let mut current_section = Section::default();
-    for (line_idx, line) in input.lines().enumerate() {
-        // Match on the first character
-        match line.chars().next() {
-            Some('[') => {
-                // This is a section
-                let end = line.find(']');
-                let name = match end {
-                    // Allow there to be a missing ] in the section header
-                    Some(idx) => line[1..idx].trim(),
-                    None => line[1..].trim(),
-                };
-                // Simultaneously push the previous section and create this new one
-                file.sections.push(std::mem::replace(
-                    &mut current_section,
-                    Section {
-                        name: Some(name),
-                        span: Span::from_line(line_idx + 1),
-                        values: Vec::default(),
-                    },
-                ));
-            }
-            Some(..) => {
-                // This is a piece of data
-                let equals = line.find('=');
-                let data = match equals {
-                    // Key Value Pair
-                    Some(idx) => {
-                        let key = line[0..idx].trim();
-                        let value = line[(idx + 1)..].trim();
-                        ValueData::KeyValuePair { key, value }
-                    }
-                    // No Equals it's a Value
-                    None => {
-                        let value = line.trim();
-                        ValueData::Value { value }
-                    }
-                };
-                // Push data onto current section
-                current_section.values.push(Value {
-                    span: Span::from_line(line_idx + 1),
-                    data,
-                });
-            }
-            // Empty line, ignore it
-            None => {}
-        }
-    }
-
-    // Push the last section into the file
-    file.sections.push(current_section);
-
-    file
+pub trait FromKVPFile {
+    fn from_kvp_file(k: &KVPFile<'_>) -> Self;
 }
 
-#[cfg(test)]
-mod test {
-    use crate::parse::kvp::{parse_kvp_file, KVPFile, Section, Value, ValueData};
-    use crate::parse::Span;
+pub trait FromKVPSection {
+    fn from_kvp_section(section: &KVPSection<'_>) -> Self;
+}
 
-    #[test]
-    fn value() {
-        let kvp = parse_kvp_file("my_value");
-        assert_eq!(
-            kvp,
-            KVPFile {
-                sections: vec![Section {
-                    name: None,
-                    span: Span::from_line(0),
-                    values: vec![Value {
-                        span: Span::from_line(1),
-                        data: ValueData::Value { value: "my_value" }
-                    }]
-                }]
-            }
-        );
-    }
-
-    #[test]
-    fn kvp() {
-        let kvp = parse_kvp_file("my_key = my_value");
-        assert_eq!(
-            kvp,
-            KVPFile {
-                sections: vec![Section {
-                    name: None,
-                    span: Span::from_line(0),
-                    values: vec![Value {
-                        span: Span::from_line(1),
-                        data: ValueData::KeyValuePair {
-                            key: "my_key",
-                            value: "my_value",
-                        }
-                    }]
-                }]
-            }
-        );
-    }
-
-    #[test]
-    fn named_section_value() {
-        let kvp = parse_kvp_file("[my_section]\nmy_value");
-        assert_eq!(
-            kvp,
-            KVPFile {
-                sections: vec![
-                    Section {
-                        name: None,
-                        span: Span::from_line(0),
-                        values: Vec::default(),
-                    },
-                    Section {
-                        name: Some("my_section"),
-                        span: Span::from_line(1),
-                        values: vec![Value {
-                            span: Span::from_line(2),
-                            data: ValueData::Value { value: "my_value" }
-                        }]
-                    }
-                ]
-            }
-        );
-    }
-
-    #[test]
-    fn named_section_kvp() {
-        let kvp = parse_kvp_file("[my_section]\nmy_key = my_value");
-        assert_eq!(
-            kvp,
-            KVPFile {
-                sections: vec![
-                    Section {
-                        name: None,
-                        span: Span::from_line(0),
-                        values: Vec::default(),
-                    },
-                    Section {
-                        name: Some("my_section"),
-                        span: Span::from_line(1),
-                        values: vec![Value {
-                            span: Span::from_line(2),
-                            data: ValueData::KeyValuePair {
-                                key: "my_key",
-                                value: "my_value",
-                            }
-                        }]
-                    }
-                ]
-            }
-        );
-    }
-
-    #[test]
-    fn empty_section_name() {
-        let kvp = parse_kvp_file("[]");
-        assert_eq!(
-            kvp,
-            KVPFile {
-                sections: vec![
-                    Section {
-                        name: None,
-                        span: Span::from_line(0),
-                        values: Vec::default(),
-                    },
-                    Section {
-                        name: Some(""),
-                        span: Span::from_line(1),
-                        values: Vec::default(),
-                    }
-                ]
-            }
-        );
-    }
-
-    #[test]
-    fn section_name_no_rbracket() {
-        let kvp = parse_kvp_file("[my_section");
-        assert_eq!(
-            kvp,
-            KVPFile {
-                sections: vec![
-                    Section {
-                        name: None,
-                        span: Span::from_line(0),
-                        values: Vec::default(),
-                    },
-                    Section {
-                        name: Some("my_section"),
-                        span: Span::from_line(1),
-                        values: Vec::default(),
-                    }
-                ]
-            }
-        );
-    }
+pub trait FromKVPData {
+    fn from_kvp_data(data: &KVPValue<'_>) -> Self;
 }
